Fixes post pagination
