+++ src/components/pages/profile/friends-post/index.js
@@ -1,0 +1,7 @@
+export const FriendsPost = () => {
+    return (
+        <div>Posts from friends here!</div>
+    )
+};
+
+export default FriendsPost;
