+++ src/components/pages/profile/friends/index.js
@@ -1,0 +1,7 @@
+export const Friends = ({ username }) => {
+    return (
+        <div>FRIENDS HERE!</div>
+    )
+};
+
+export default Friends;
